backend/logbook/serializers.py
from rest_framework import serializers
from .models import DriverLog, LogEntry

class LogEntrySerializer(serializers.ModelSerializer):
    class Meta:
        model = LogEntry
        fields = '__all__'

class DriverLogSerializer(serializers.ModelSerializer):
    entries = LogEntrySerializer(many=True, read_only=True)

    class Meta:
        model = DriverLog
        fields = '__all__'