backend/requirements.txt
Django
gunicorn
djangorestframework
requests
